master control socket errors are not fatal.

+++ zygote/master.py
@@ -114,21 +114,26 @@ class ZygoteMaster(object):
         self.io_loop.add_handler(self.master_socket.fileno(), self.handle_protocol_msg, self.io_loop.READ)
 
     def setup_control_socket(self):
-        socket_path = self.control_socket_path
-        if os.path.exists(socket_path):
-            # NOTE: Starting the same application twice we won't get
-            # here since main() won't be able to bind. We can add a
-            # (ex|nb) file lock if needed.
-            log.error("Control socket exitsts %s. Probably from a previous run. Removing...", socket_path)
-            self.cleanup_control_socket()
-        log.debug("Binding to control socket %s", socket_path)
-        self.control_socket = socket.socket(socket.AF_UNIX, socket.SOCK_DGRAM, 0)
-        self.control_socket.bind(socket_path)
-        self.io_loop.add_handler(self.control_socket.fileno(), self.handle_control_msg, self.io_loop.READ)
+        try:
+            socket_path = self.control_socket_path
+            if os.path.exists(socket_path):
+                # NOTE: Starting the same application twice we won't get
+                # here since main() won't be able to bind. We can add a
+                # (ex|nb) file lock if needed.
+                log.error("Control socket exitsts %s. Probably from a previous run. Removing...", socket_path)
+                self.cleanup_control_socket()
+            log.debug("Binding to control socket %s", socket_path)
+            self.control_socket = socket.socket(socket.AF_UNIX, socket.SOCK_DGRAM, 0)
+            self.control_socket.bind(socket_path)
+            self.io_loop.add_handler(self.control_socket.fileno(), self.handle_control_msg, self.io_loop.READ)
+        except Exception, e:
+            log.error("Can not bind to control socket: %s", e)
+            log.warning("Zygote is starting but you will not be able to make configuration changes on the running zygote master.")
 
     def cleanup_control_socket(self):
-        log.info("Removing %s" % self.control_socket_path)
-        os.unlink(self.control_socket_path)
+        if os.path.isfile(self.control_socket_path):
+            log.info("Removing %s" % self.control_socket_path)
+            os.unlink(self.control_socket_path)
 
     def handle_control_msg(self, fd, events):
         assert fd == self.control_socket.fileno()
